+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-1.0.2
+1.0.4
+
+## 1.0.4
+## 2020-04
+- Fix an error that dataloader may mess up the file order
+- Add a dataset checker to help verify the DDF 
+
+## 1.0.3
+## 2020-02
+- Add --export to Train/eval.py
 
 ## 1.0.2
 ## 2020-02-16